Add initial MkDocs configuration and documentation
This commit introduces the initial setup for MkDocs. It includes the creation of the configuration file (mkdocs.yml) with the site name, URL, copyright, repo URL and name, theme, and color palette. Furthermore, it adds an "index.md" file in the docs folder, providing basic information about the project and documentation commands.