Add new 'data' module and update service configurations
Introduced a new 'data' module in the existing service stack and generated relevant files for the module including resource files and main class. Furthermore, made updates to server port and database settings within the 'user' microservice. These changes extend the capabilities of the application by adding a new functionality and enhance the configuration of existing services.