Refactor Weather class and remove unnecessary properties
Significant simplification of the Weather class properties was made by removing unnecessary parameters that were overloading the class. Also, a slight naming adjustment was made in the WeatherTasks class to enhance clarity in line with best practices. This should considerably streamline the weather information processing and improve the code's readability and maintainability.