Add battleDeaths to War entity and improve war data import
Added a new field 'battleDeaths' to the War entity and implemented an improved function for importing war data in the COWTasks class. Also introduced a new outcome type 'NONE' in the WarOutcome enum for better handling of different war outcomes. These changes enhance data accuracy and comprehensiveness for war-related information.