Add new application properties and update data conversions
The code updates introduce new properties to the application config files for starting specific job types. In addition, a new utility function is implemented for converting string values to integers. This function is then integrated in the relevant parts of the application to handle potential NumberFormatExceptions. A datatype in the `Religion` entity class is also modified for more accurate data representation.