Add TerritorialChangeTask and update dependencies
A new task, TerritorialChangeTask, has been added to handle territory change events in the COWTasks file. It processes territory change data from an unpacked ".zip" file and persists the data into the database. In addition, the versions of some dependencies in the pom.xml have been updated to be more current.