Add Grafana deployment to monitoring stack
This commit introduces the deployment file for Grafana in the monitoring stack within the 'livius' namespace. It also changes the container port and the service port of Tempo from 3100 to 3200. This update is crucial as it allows Grafana to receive data from various sources, including Prometheus and Tempo, thus improving observability.