Add Trade entity and repository, update dependencies
Added a new Trade entity along with its repository implementation. This new entity represents a trade document in a MongoDB database and contains various properties to describe the characteristics of a trade. The entity is accompanied by its corresponding repository for storing and retrieving trade data. The dependencies versions of entities and repositories in the pom.xml files of the projects have also been updated.