Add info.yaml for deployment and service configuration
A new configuration file, info.yaml, was added to the backend deployment folder. It contains the necessary configurations for deploying an API ("info") within Kubernetes, including the specifications of the replica, strategy, template, and the corresponding services. It also has the environment variables pertaining to EUREKA, the database and LOKI URI configured.