Expand Newsletter entity with new attributes
The 'Newsletter' class has been extended with additional fields: 'subject', 'recipients', 'data', and 'template'. These new attributes enrich the information model representing a Newsletter. The 'data' variable is a HashMap used to store key-value pairs of additional dynamic data. The 'template' field is designed to hold the content of the newsletter.