Update documentation for Kubernetes in Livius project
The documentation for Kubernetes in the Livius project has been significantly expanded. It now includes a more detailed explanation on why Kubernetes is used in the project, the location of all Kubernetes configs, and information on how to deploy Livius in individual infrastructures. Special emphasis is made on using ArgoCD for deployments.