Update UserCard display in adaptive layout
Updated the home screen to utilize adaptive layout for different screen sizes. In the medium and larger versions, the UserCard display now incorporates a Row and Column structure. This change also improves the handling of cases with no user data, replacing a previous CircularProgressIndicator with a more informative text message.