Add User, Role, and CommandSettings classes
New classes have been added to the Entities and Commands namespaces. These include User and Role classes for the User namespace and CommandSettings class for the Commands namespace. The inclusion of the Commands and Entities folders in the project file has been removed, as they are now implicitly included by their respective files.